+++ packages/ui/src/listen/minimalism/music-widget/index.ts
@@ -1,3 +1,3 @@
-import MusicWidget from './MusicWidget';
+import MusicWidget from './music-widget';
 
 export default MusicWidget;
